Move poetryLocator.ts
